# src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fingerprint.AI Gateway API
  version: 1.1.0
  description: >-
    Web2.5 Gateway API for the AI Agent Fingerprinting System. Provides agent
    registry lookups, behavioral drift verification, clinical audit event
    logging, triage encounter management, Merkle-tree anchoring to
    Ethereum/Sepolia, and cryptographic integrity health checks.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
  /v1/agents:
    get:
      tags:
        - Agents
      summary: List registered AI agents
      description: >-
        Returns a paginated list of all globally registered agents. Supports a
        limit parameter; cursor-based pagination is reserved for future use.
      operationId: listAgents
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of agents to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated list of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/agents/{fingerprintHash}:
    get:
      tags:
        - Agents
      summary: Get agent profile by fingerprint hash
      description: >-
        Fetches a detailed agent profile. Results are cached in Redis for
        low-latency responses; cache misses fall through to PostgreSQL.
      operationId: getAgentProfile
      security:
        - BearerAuth: []
      parameters:
        - name: fingerprintHash
          in: path
          required: true
          description: The keccak256 fingerprint hash of the registered agent.
          schema:
            type: string
            pattern: '^0x[a-fA-F0-9]{64}$'
      responses:
        '200':
          description: Detailed agent profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentProfile'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: No agent found for the provided fingerprint hash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/agents/verify:
    post:
      tags:
        - Verification
      summary: Verify an AI agent's behavioral fingerprint
      description: >-
        The core verification endpoint for relying parties. Accepts a fingerprint
        hash and an optional current ResponseSet (behavioral trait payload), then
        runs a multi-signal trust evaluation including revocation checks,
        Jaccard similarity scoring, perturbation analysis, and contextual rules.
        Returns a trust decision of accept, challenge, or deny.
      operationId: verifyAgentFingerprint
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyAgentRequest'
      responses:
        '200':
          description: Verification result with trust score and decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyAgentResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: No agent found for the provided fingerprint hash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Baseline responses not synced to off-chain cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/internal/traits/seed:
    post:
      tags:
        - Internal
      summary: Seed behavioral trait baseline into off-chain cache
      description: >-
        Internal development endpoint. Stores a raw behavioral ResponseSet in
        Redis keyed by fingerprint hash, mimicking IPFS sidecar delivery used
        during agent registration. The endpoint verifies that the ResponseSet
        produces the expected trait hash before persisting.
      operationId: seedBehavioralTraits
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeedTraitsRequest'
      responses:
        '200':
          description: Traits successfully seeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeedTraitsResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/events:
    post:
      tags:
        - Events
      summary: Append an immutable clinical event to the audit log
      description: >-
        Ingests a new clinical interaction into the append-only audit ledger.
        Validates that reference pointers conform to hash or URI formats (PHI
        guardrail). Each event is hash-chained to its predecessor for
        cryptographic integrity.
      operationId: createClinicalEvent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event successfully appended to the ledger
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEventResponse'
        '400':
          description: Missing required fields or PHI violation risk
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      tags:
        - Events
      summary: Fetch clinical audit events
      description: >-
        Retrieves the immutable clinical audit log with optional filters by
        agent and time range. Returns the most recent 100 events.
      operationId: listClinicalEvents
      security:
        - BearerAuth: []
      parameters:
        - name: agent_fingerprint_id
          in: query
          required: false
          description: Filter events by agent fingerprint hash.
          schema:
            type: string
        - name: days_back
          in: query
          required: false
          description: Number of days to look back from now.
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: List of clinical audit events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/events/anchor/trigger:
    post:
      tags:
        - Anchoring
      summary: Trigger Merkle-tree anchoring of pending events
      description: >-
        Development trigger that forces the background anchoring job to run
        immediately. Builds a Merkle tree from all unanchored event hashes,
        simulates a Sepolia smart contract call, and marks the events as
        anchored. Events with invalid hashes are quarantined.
      operationId: triggerEventAnchoring
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Anchoring result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnchorResult'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v1/triage/encounters/{session_id}/action:
    post:
      tags:
        - Triage
      summary: Log a clinician decision on an encounter
      description: >-
        Records a clinician's accept, override, downgrade, or escalation
        decision against an existing triage encounter. If a decision already
        exists for the encounter, the new entry is treated as an amendment
        with provenance linkage.
      operationId: logClinicianAction
      security:
        - BearerAuth: []
      parameters:
        - name: session_id
          in: path
          required: true
          description: The session identifier for the encounter.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClinicianActionRequest'
      responses:
        '200':
          description: Clinician action logged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicianActionResponse'
        '400':
          description: Invalid or missing action value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Active triage agent is missing or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >-
        API key-based authentication. Include the API key as a Bearer token
        in the Authorization header.
  schemas:
...
    ApiError:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
              example: internal_error
            message:
              type: string
              description: Human-readable error description.
              example: An internal error occurred
            details:
              type: object
              description: Additional context for the error.
              additionalProperties: true

    # ── Agents ─────────────────────────────────────────────────────────────
...
    AgentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AgentListItem'
        has_more:
          type: boolean
          description: Whether more results are available beyond the current page.
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for fetching the next page (reserved for future use).
...
    AgentProfile:
      type: object
      properties:
        fingerprintHash:
          type: string
        agent_id:
          type: string
          description: Logical slug identifier used for role resolution.
        name:
          type: string
        provider:
          type: string
        version:
          type: string
        registeredBy:
          type: string
          description: Ethereum address of the registrant.
        createdAt:
          type: string
          format: date-time
        isRevoked:
          type: boolean
        behavioralTrait:
          $ref: '#/components/schemas/BehavioralTraitInfo'
        indexer:
          $ref: '#/components/schemas/IndexerHealth'

    # ── Verification ───────────────────────────────────────────────────────
...